move a bunch of files to/from absint/
Summary:
This is needed to make absint/ its own dune library. Modules in absint/
cannot depend on modules in backend/. This is because backend/Summary.ml
depends on *all* the (inteprocedural) analyses since its stores elements
of their domains. But analyses are supposed to depend on absint/ to
use the AI framework and others. Thus absint/ cannot depend on backend/
otherwise it creates a dependency cycle.

Right now infer builds with this cycle because we consider all these
directories to be one library and there is no dependency cycle between
the individual modules, but making absint/ a library makes the situation
coarser.

One actual change in there: one function in PatternMatch was specific to
nullsafe and was moved there.

Reviewed By: ngorogiannis

Differential Revision: D21257467

fbshipit-source-id: 42af43315