[nullsafe] Aggregate meta-issues for top-level classes
Summary:
This diff changes way we treat classes w.r.t. to Nullsafe modes when
issuing meta-issues.

Previously, we considered nested class independently of the outer one.
This was leading to a tricky case: when the class is clean but nested
class needs fixing, meta-info told that class can be Nullsafe.

This is counter-intutive and lead to problems when users tried to follow
wrong nullsafe suggestions for this case.

After this diff we:
1. Start calculating meta-issues only on the outermost level. This will simplify
reasoning about nullsafe stats.
2. Aggregate all nested issues counters to corresponding outer-level class.

Among others, CLASS_CAN_BE_NULLSAFE Advice will finally become
actionable in all known cases.

Reviewed By: artempyanykh

Differential Revision: D21353607

fbshipit-source-id: a17c6958a