[process] redirect desired output stream to file, while logging the other one
Summary: Build tool wrappers sometimes have issues when their output streams are redirected to those of infer's while one of those streams is read and used in infer.  This diff side-steps this problem by always redirecting the stream of interest to a file (which also helps in debugging), while logging the other stream through `Logging.progress`.

Reviewed By: jvillard

Differential Revision: D21404736

fbshipit-source-id: 04c92799c