[pulse] Model CFRelease as removing the Allocated attribute rather than as free
Summary:
Because in the real semantics CFRelease can be used more than once, and also the variables can be used after CFRelease in general, modelling this as `free` causes many `USE_AFTER_FREE` errors. Now we change the model to not add the `Invalid CFree` attribute, but to just remove the `Allocated` attribute. So we can model memory leaks in the simple case of `Create` and not `CFRelease` before going out of scope, but we avoid the `USE_AFTER_FREE`.

Since the model for CFRelease now diverges from free, changed the command line option for modelling to `pulse-model-release-pattern`.

Reviewed By: jvillard

Differential Revision: D21324895

fbshipit-source-id: ab323d981