[frontend][clang][RFC] fix translation of enum constant
Summary:
Previous translation of enum constants were wrong since they assumed that the enum constant didn't include any global variable (hence they just looked up the enum exp from the map, forgetting to tie the respective instructions into the cfg).

```
const int gvar = 0;

enum {
  evar = gvar,

};

 int dangling() {
  return evar;
}
```

as a result, the CFG was missing the instruction for the load of the `gvar`.

{F237004587}

This diff fixes this issue by hooking up the instructions that load the enum constant in to the CFG. Note that in this example, it is only a load instruction but there could be more instructions (e.g. if we had `gvar > 1`, we would have prune +join).

{F237004493}

Reviewed By: ezgicicek

Differential Revision: D21549781

fbshipit-source-id: 525534fb2