[control] Fix dangling node
Summary:
Consider the below program,

```
const int gvar = 0;
enum {
  cvar = gvar + 1,

};

bool dangling_cvar(int x) {
  for (int i = 0; i < cvar; i++){
  }
}
```
In the prune node, we don't have `cvar` but its inlined version, i.e. we have  ` i < n$2 + 1` and the variable `n$2` is defined not in predecessor nodes to the prune node (as normally one would expect) but in a separate dangling node (see the {F236910156})

:

```
6: BinaryOperatorStmt:Add n$2=*&#GB<test.cpp|ice>$gvar:int [line 38, column 10]
```
When computing the control var of this loop, previously we gave up and simply raised an error.

With this diff, let's handle this case by looking inside this dangling node.

Reviewed By: skcho

Differential Revision: D21525569

fbshipit-source-id: bd4371493