[cost] Add ability to test costs-report.json
Summary:
In order to test cost analysis results, currently we rely on having an arbitrary cost threshold (200) and report issues that exceed this cost. For instance, a cost of 201 is considered expensive and reported as `EXPENSIVE_EXECUTION_TIME` issue in cost tests.

This means, if we change the cost analysis in a slight way that results in some constant cost increase under 200, we wouldn't able to detect it. I find this unsatisfactory and somewhat hacky.

This diff adds the ability to write the result of `costs-report.json` into a separate `cost-issues.exp`  and then compare the actual costs (not only than relying on this arbitrary threshold reporting mechanism).

Reviewed By: skcho

Differential Revision: D21816312

fbshipit-source-id: 93b531928