[cost] Remove cost threshold from tests
Summary:
Now that we have a way to write cost issues, let's not rely on some arbitrary threshold (and also get rid of `EXPENSIVE_EXECUTION_TIME` issues in tests).

One consequence of this is that we will loose the cost traces in tests since `costs-report.json` doesn't have any traces.  Next diff fixes that.

Reviewed By: skcho

Differential Revision: D21837574

fbshipit-source-id: 86b4d028d