[starvation] fix static field printing
Summary:
The Java frontend translates java field accesses on an object reference as field (`.`) followed by dereference (`*`) exactly like the arrow operator in C.  However, when the field is static, it generates just a field access `.`.  This is a bug.

This diff mitigates its effects for printing locks in starvation.

Reviewed By: skcho

Differential Revision: D21882875

fbshipit-source-id: 79846d826