Checker.name -> Checker.id
Summary:
The checker names are only used in debug information but I need them to
be more useful so users can do queries about each checker. Turn them
into an "id" instead of a "name", with some constraints to avoid crazy
IDs. In the next diff these IDs will be used on the command lin.

Reviewed By: dulmarod

Differential Revision: D21934373

fbshipit-source-id: 847a4958d