skeleton new `infer help` command
Summary:
`infer help` will be used to display information about issue types and
checkers, and to generate the corresponding website documentation. We
can add more things in it over time. The goal is to avoid having to go
read the source code of infer to figure things out that are user-facing.

Reviewed By: ezgicicek

Differential Revision: D21934376

fbshipit-source-id: 2788c5af1