[buck-clang] don't special case for targets
Summary:
There is a lot of subtlety in our parsing of buck targets on the command line, that is then just thrown away. Push this one level up, getting rid of the special case where in Clang mode if we only have "normal" targets we don't resolve them.

Also introduce a proper variant for buck target types.

Reviewed By: skcho

Differential Revision: D22160490

fbshipit-source-id: 500c1b12c