[infer] NULLPTR_DEREFERENCE false positive caused by thread_local variable
Summary:
Keyword `thread_local` in cpp allows us to create a variable with thread storage duration, meaning that the object's lifetime begins when the thread begins and ends when the thread ends.

We get `NULLPTR_DEREFERENCE` false positive for `thread_local` variable since we reallocate it in the `VariableLifetimeBegins` metadata instruction and we do not see further updates to the variable. To solve the issue we special case `VariableLifetimeBegins`  instruction for global variables.

Reviewed By: jvillard

Differential Revision: D22284135

fbshipit-source-id: 13c14ef90