[cost] add model for NSString.string_with_utf8_string_linear and NSString.length
Summary:
As title
Model `NSString` as `JavaString`.
Since `NSArray` does not contain information about its type of element, we do not use associate string with collection as in Java and C++. In Java, String model is implemented using java collection, and for C++, string model is implemented using vector.
So instead, we use existing `JavaString` model.

Reviewed By: skcho

Differential Revision: D22431949

fbshipit-source-id: 7cdde1ad7